feito compnente interno do card
